+++ src/app/wines/Components/CardWine/CardWine.tsx
